refactor: move state retated to state.rs
